% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_scrap.R
\name{colors_scrap}
\alias{colors_scrap}
\title{Scrape Color Codes from a Web Page}
\usage{
colors_scrap(link, askRobot = FALSE)
}
\arguments{
\item{link}{the link of the web page to scrape. Can be a character vector
of multiple URLs.}

\item{askRobot}{logical. Should the function ask the robots.txt if we're
allowed or not to scrape the web page? Default is \code{FALSE}.}
}
\value{
a character vector of unique color codes found on the page, or
\code{invisible(NULL)} if none are found, or \code{NA} on error.
}
\description{
This function scrapes all color codes present within a given
website. It searches inline \code{style} attributes, \code{<style>} tags,
and linked external CSS stylesheets for color values in the following
formats: hexadecimal (\code{#RGB}, \code{#RRGGBB}, \code{#RRGGBBAA}),
\code{rgb()}, \code{rgba()}, \code{hsl()}, and \code{hsla()}.
}
\examples{
\donttest{

colors_scrap(link = "https://ropensci.org/")

}

}
