% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_scrap.R
\name{csv_scrap}
\alias{csv_scrap}
\title{Scrape and download CSV files from a Web Page}
\usage{
csv_scrap(link, path = getwd(), askRobot = FALSE)
}
\arguments{
\item{link}{the link of the web page}

\item{path}{the path where to save the CSV files. Defaults to the current directory}

\item{askRobot}{logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE.}
}
\value{
called for the side effect of downloading CSV files from a website
}
\description{
Scrape and download CSV files from a Web Page
}
