% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_noalt_scrap.R
\name{images_noalt_scrap}
\alias{images_noalt_scrap}
\title{Scrape Images URLS that don't have 'alt' attributes}
\usage{
images_noalt_scrap(link, askRobot = FALSE)
}
\arguments{
\item{link}{the URL of the web page}

\item{askRobot}{logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE.}
}
\value{
a character vector of images' URL without "alt" attribute
}
\description{
Scrape Images URLS that don't have 'alt' attributes
}
\examples{
\donttest{

images_noalt_scrap(link = "https://www.r-consortium.org/")

}

}
