% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_images.R
\name{images_preview}
\alias{images_preview}
\title{Scrape Images URLs}
\usage{
images_preview(link, askRobot = FALSE)
}
\arguments{
\item{link}{the link of the web page}

\item{askRobot}{logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE.}
}
\value{
Images URLs
}
\description{
Scrape Images URLs
}
\examples{
\donttest{

images_preview(link = "https://rstudio.com/")

}

}
