% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images_scrap.R
\name{images_scrap}
\alias{images_scrap}
\title{Scrape Images from a Web Page}
\usage{
images_scrap(link, imgpath = getwd(), extn, askRobot = FALSE)
}
\arguments{
\item{link}{the link of the web page}

\item{imgpath}{the path of the images. Defaults to the current directory}

\item{extn}{the extension of the image: png, jpeg ...}

\item{askRobot}{logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE.}
}
\value{
Images
}
\description{
Scrape Images from a Web Page
}
\examples{
\dontrun{

images_scrap(link = "https://rstudio.com/", extn = "png")

}

}
