% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_img_no_alt.R
\name{img_no_alt_scrap}
\alias{img_no_alt_scrap}
\title{Scrape Images URLS that don't have 'alt' attributes}
\usage{
img_no_alt_scrap(link, askRobot = FALSE)
}
\arguments{
\item{link}{the URL of the web page}

\item{askRobot}{}
}
\value{

}
\description{
Scrape Images URLS that don't have 'alt' attributes
}
