% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_noalt_scrap.R
\name{img_noalt_scrap}
\alias{img_noalt_scrap}
\title{Scrape Images URLS that don't have 'alt' attributes}
\usage{
img_noalt_scrap(link, askRobot = FALSE)
}
\arguments{
\item{link}{the URL of the web page}

\item{askRobot}{}
}
\value{

}
\description{
Scrape Images URLS that don't have 'alt' attributes
}
