% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_scrap.R
\name{pdf_scrap}
\alias{pdf_scrap}
\title{Scrape PDF documents from a Web Page}
\usage{
pdf_scrap(link, pdfname = "pdf", pdfpath = getwd(), askRobot = FALSE)
}
\arguments{
\item{link}{the link of the web page}

\item{pdfname}{how the PDF documents should be named. Defaults to "pdf" which will yield pdf-1, pdf-2 ... etc.}

\item{pdfpath}{the path of the PDF files. Defaults to the current directory}

\item{askRobot}{logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE.}
}
\value{
PDF documents
}
\description{
Scrape PDF documents from a Web Page
}
\examples{
\donttest{

 pdf_scrap(link = "https://www.worldbank.org/en/research",
           pdfname = "amazing_pdf",
           pdfpath = getwd())
}

}
