% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap.R
\name{scrap}
\alias{scrap}
\title{function to scrap a website}
\usage{
scrap(link, node)
}
\arguments{
\item{link}{the link of the webpage to scrap}

\item{node}{the CSS element to consider, the SelectorGadget tool is highly recommended}
}
\value{
a character vector
}
\description{
function to scrap a website
}
\examples{
# Extracting the books' title of the 20 Best Sellers of 2019 from the barnesandnoble website
donttest{tidy_scrap("https://www.barnesandnoble.com/b/the-top-100-bestsellers-of-the-year/_/N-1p4d", ".product-info-title a")}
}
