% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_three.R, R/scrap_two.R
\name{scrap_df}
\alias{scrap_df}
\title{function to build a three columns dataframe through scraping}
\usage{
scrap_df(link, nodes)

scrap_df(link, nodes)
}
\arguments{
\item{link}{the link of the webpage to scrap}

\item{nodes}{the CSS elements to consider, the SelectorGadget tool is highly recommanded}

\item{link}{the link of the webpage to scrap}

\item{nodes}{the CSS elements to consider, the SelectorGadget tool is highly recommanded}
}
\description{
function to build a three columns dataframe through scraping

function to build a two columns dataframe through scraping
}
