% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_more.R
\name{scrap_more}
\alias{scrap_more}
\title{Scraping multiple pages of a website}
\usage{
scrap_more(link, node, from = 0, to = 20)
}
\arguments{
\item{link}{the link of the webpage to scrap}

\item{node}{the CSS element to consider, the SelectorGadget tool is highly recommanded}

\item{from}{the first page to begin scraping, default equal 0}

\item{to}{the last page of scraping, default equal 20}
}
\description{
Scraping multiple pages of a website
}
