% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_this.R
\name{scrap_this}
\alias{scrap_this}
\title{function to scrap a website}
\usage{
scrap_this(link, node)
}
\arguments{
\item{link}{the link of the webpage to scrap}

\item{node}{the CSS element to consider, the SelectorGadget tool is highly recommanded}
}
\description{
function to scrap a website
}
