% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_scrap.R
\name{table_scrap}
\alias{table_scrap}
\title{HTML table scraping}
\usage{
table_scrap(link, askRobot = FALSE)
}
\arguments{
\item{link}{the link of the webpage containing the table to scrap}

\item{askRobot}{logical. Should the function ask the robots.txt if we're allowed or not to scrap the web page ? Default is FALSE.}
}
\value{
a tidy dataframe.
}
\description{
This function is used to scrap an html table from a website.
}
\examples{
\donttest{
# Extracting premier ligue 2019/2020 top scorers

link     <- "https://www.topscorersfootball.com/premier-league"
table_scrap(link)}

}
