% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scrap.R
\name{tidy_scrap}
\alias{tidy_scrap}
\title{function to scrap a tidy dataframe from a website}
\usage{
tidy_scrap(link, nodes, scrap_names)
}
\arguments{
\item{link}{the link of the webpage to scrap}

\item{nodes}{the vector of CSS elements to consider, the SelectorGadget tool is highly recommended}

\item{scrap_names}{the names of the expected columns}
}
\value{
a tidy dataframe
}
\description{
function to scrap a tidy dataframe from a website
}
