% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titles_scrap.R
\name{titles_scrap}
\alias{titles_scrap}
\title{Website title scraping}
\usage{
titles_scrap(link, contain = NULL, askRobot = FALSE)
}
\arguments{
\item{link}{the link of the web page to scrape}

\item{contain}{filter the titles according to the character string provided.}

\item{askRobot}{logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE}
}
\value{
a character vector
}
\description{
This function is used to scrape titles from a website. Useful for scraping daily electronic newspapers' titles.
}
\examples{
\donttest{
# Extracting the current titles of the New York Times

link     <- "https://www.nytimes.com/"

titles_scrap(link)}

}
