% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_scrap.R
\name{colors_scrap}
\alias{colors_scrap}
\title{Scrape Color Codes from a Web Page}
\usage{
colors_scrap(link, askRobot = FALSE)
}
\arguments{
\item{link}{the link of the web page to scrape}

\item{askRobot}{logical. Should the function ask the robots.txt if we're
allowed or not to scrape the web page? Default is \code{FALSE}.}
}
\value{
a character vector of unique color codes found on the page, or
\code{NULL} if none are found.
}
\description{
This function is used to extract all color codes present within
a given website. It searches inline \code{style} attributes and
\code{<style>} tags for color values in \code{#Hex}, \code{rgb()},
\code{rgba()}, \code{hsl()}, and \code{hsla()} formats.
}
\examples{
\dontrun{

colors_scrap(link = "https://www.r-project.org/")

}

}
