% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap.R
\name{scrap}
\alias{scrap}
\title{function to scrap a website}
\usage{
scrap(link, node)
}
\arguments{
\item{link}{the link of the webpage to scrap}

\item{node}{the CSS element to consider, the SelectorGadget tool is highly recommended}
}
\value{
a character vector
}
\description{
function to scrap a website
}
\examples{
\donttest{
# Extracting imdb top 250 movie titles

link <- "https://www.imdb.com/chart/top/"
node <- ".titleColumn a"

tidy_scrap(link, node)}

}
