% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap.R
\name{scrap}
\alias{scrap}
\title{Simple website scraping}
\usage{
scrap(link, node, clean = FALSE, askRobot = FALSE)
}
\arguments{
\item{link}{the link of the webpage to scrap}

\item{node}{the HTML or CSS element to consider, the SelectorGadget tool is highly recommended}

\item{clean}{logical. Should the function clean the extracted vector or not ? Default is FALSE.}

\item{askRobot}{logical. Should the function ask the robots.txt if we're allowed or not to scrap the web page ? Default is FALSE.}
}
\value{
a character vector
}
\description{
This function is used to scrap one element from a website.
}
\examples{
\donttest{
# Extracting imdb top 250 movie titles

link <- "https://www.imdb.com/chart/top/"
node <- ".titleColumn a"

scrap(link, node)}

}
