% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_scrap.R
\name{table_scrap}
\alias{table_scrap}
\title{HTML table scraping}
\usage{
table_scrap(link, choose = 1, header = T, askRobot = FALSE)
}
\arguments{
\item{link}{the link of the web page containing the table to scrape}

\item{choose}{an integer indicating which table to scrape}

\item{header}{do you want the first line to be the leader (default to TRUE)}

\item{askRobot}{logical. Should the function ask the robots.txt if we're allowed or not to scrape the web page ? Default is FALSE.}
}
\value{
a data frame object.
}
\description{
This function is used to scrape an html table from a website.
}
\examples{
\donttest{
# Extracting premier ligue 2019/2020 top scorers

link     <- "https://www.topscorersfootball.com/premier-league"
table_scrap(link)}

}
