% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scrap.R
\name{tidy_scrap}
\alias{tidy_scrap}
\title{Tidy scrapping from a website}
\usage{
tidy_scrap(link, nodes, scrap_names)
}
\arguments{
\item{link}{the link of the webpage to scrap}

\item{nodes}{the vector of CSS elements to consider, the SelectorGadget tool is highly recommended}

\item{scrap_names}{the names of the expected columns}
}
\value{
a tidy dataframe
}
\description{
This function is used to scrap a tibble from a website.
}
\examples{
\donttest{
# Extracting imdb movie titles and rating

link     <- "https://www.imdb.com/chart/top/"
my_nodes <- c(".titleColumn a", "strong")
names    <- c("title", "rating")

tidy_scrap(link, my_nodes, names)}

}
