% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_scrap.R
\name{tidy_scrap}
\alias{tidy_scrap}
\title{function to scrap a tidy dataframe from a website}
\usage{
tidy_scrap(link, nodes, scrap_names)
}
\arguments{
\item{link}{the link of the webpage to scrap}

\item{nodes}{the vector of CSS elements to consider, the SelectorGadget tool is highly recommended}

\item{scrap_names}{the names of the expected columns}
}
\value{
a tidy dataframe
}
\description{
function to scrap a tidy dataframe from a website
}
\examples{
# Extracting as a tidy dataframe the books' title and author of the 20 Best Sellers of 2019 from the barnesandnoble website
\donttest{
tidy_scrap("https://www.barnesandnoble.com/b/the-top-100-bestsellers-of-the-year/_/N-1p4d",
   c(".product-info-title a", ".contributors a"),
   c("title", "author"))
}
}
